Move PointsAPI to utils package
